# docs-site/api/openapi.json
{
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
...
  "info": {
    "contact": {
      "name": "tscrond",
      "url": "https://github.com/tscrond"
    },
    "description": "Fluxsend API endpoints",
    "license": {
      "name": "MIT",
      "url": "https://mit-license.org/"
    },
    "title": "FluxSend Developer API",
    "version": "1.7.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/workspaces/{workspace_id}/files/download": {
      "get": {
        "description": "Creates a signed download URL for a workspace file and redirects or proxies based on mode.",
        "parameters": [
          {
            "description": "Workspace ID",
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "File ID",
            "in": "query",
            "name": "file_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "inline or download",
            "in": "query",
            "name": "mode",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to the signed URL"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "summary": "Download workspace file",
        "tags": [
          "Workspace Files"
        ]
      }
    },
...
    "/api/workspaces/{workspace_id}/files/mkdir": {
      "post": {
        "description": "Creates a new folder inside the selected workspace.",
        "parameters": [
          {
            "description": "Workspace ID",
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "description": "Folder creation request",
          "required": true,
          "x-originalParamName": "request"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "Forbidden"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "summary": "Create workspace folder",
        "tags": [
          "Workspace Files"
        ]
      }
    },
...
    "/api/workspaces/{workspace_id}/files/tree": {
      "get": {
        "description": "Returns the file tree for a workspace path visible to the caller.",
        "parameters": [
          {
            "description": "Workspace ID",
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Path prefix",
            "in": "query",
            "name": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "Forbidden"
          }
        },
        "summary": "Get workspace file tree",
        "tags": [
          "Workspace Files"
        ]
      }
    },
    "/api/workspaces/{workspace_id}/files/upload": {
      "post": {
        "description": "Uploads a file into a workspace folder for users with write access.",
        "parameters": [
          {
            "description": "Workspace ID",
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "description": "File to upload",
                    "format": "binary",
                    "type": "string",
                    "x-formData-name": "file"
                  },
                  "folder": {
                    "description": "Destination folder",
                    "type": "string",
                    "x-formData-name": "folder"
                  }
                },
                "required": [
                  "file"
                ],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "Forbidden"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "summary": "Upload workspace file",
        "tags": [
          "Workspace Files"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "https://fluxsend.win/"
    }
  ]
}